Group 'kopia blob' commands in single file
